Add feature switch for the code
