+++ src/main/java/org/opentripplanner/util/OTPFeature.java
@@ -78,7 +78,12 @@ public enum OTPFeature {
   SandboxAPITransmodelApi(false, true, "Enable Entur Transmodel(NeTEx) GraphQL API."),
   SandboxAPITravelTime(false, true, "Enable the isochrone/travel time surface API."),
   TransferAnalyzer(false, true, "Analyze transfers during graph build."),
-  VehicleToStopHeuristics(false, true, "Enable improved heuristic for park-and-ride queries.");
+  VehicleToStopHeuristics(false, true, "Enable improved heuristic for park-and-ride queries."),
+  RealtimeResolver(
+    false,
+    true,
+    "When routing with ignoreRealtimeUpdates=true, add an extra step which populates results with realtime data"
+  );
 
   private static final Object TEST_LOCK = new Object();
